apps/cli/pyproject.toml
[project]
name = "parseltongue-cli"
version = "0.1.0"
description = "Pipeline CLI: scrape, enunciate, speak"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "parseltongue-logger",
    "parseltongue-scraper",
    "parseltongue-director",
    "python-dotenv>=1.0.0",
]

[project.scripts]
parseltongue = "parseltongue_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
parseltongue-logger = { workspace = true }
parseltongue-scraper = { workspace = true }
parseltongue-director = { workspace = true }
